Create content_check.py
This module works through URL's listed in the 'link' column of the bronze table. Using BeautifulSoup, it will determine whether or not an active site has images or links (or both), editing each row with a binary value in the 'images' and 'Ext_Links' columns. 

Since this can be a slow process, I've added a progress bar to indicate progress through the rows of the 'bronze' table.